Refactored the traits to hide the underlying types
