+++ Cargo.toml
@@ -11,11 +11,14 @@ repository = "https://github.com/jackmott/simdeez"
 categories = ["hardware-support", "science", "game-engines"]
 edition = "2018"
 
+[lib]
+doctest = false
 
 [features]
-default = ["no_std"]
+default = []
 sleef = ["sleef-sys"]
 no_std = ["dep:libm"]
+unsafe_inner_access = []
 
 [dependencies]
 paste = "1.0.11"
@@ -26,18 +29,5 @@ libm = { version = "0.2.6", optional = true }
 
 
 [dev-dependencies]
-criterion = "0.4.0"
 rand = "0.8.5"
 rand_chacha = "0.3.1"
-
-[[bench]]
-name = "sseby2"
-harness = false
-
-[[bench]]
-name = "sum"
-harness = false
-
-[[bench]]
-name = "sum2"
-harness = false
