+++ src/engines/avx2/simd.rs
@@ -1,5 +1,6 @@
 #![allow(deprecated)]
 
+use crate::Simd;
 use super::*;
 
 pub struct Avx2;
