+++ src/engines/scalar/mod.rs
@@ -1,25 +1,17 @@
-use super::*;
+use crate::{
+    libm_ext::FloatExt, InternalSimdBaseIo, SimdBaseOps, SimdConsts, SimdFloat, SimdFloat32,
+    SimdFloat64, SimdInt, SimdInt16, SimdInt32, SimdInt64,
+};
+
+use core::ops::*;
 
 mod simd;
-pub use self::overloads::*;
 pub use self::simd::*;
 
-use crate::libm_ext::FloatExt;
-
-// Newtypes for i16 vectors
-// We have to do this to allow for overloading of
-// __m128i etc
-#[derive(Copy, Clone)]
-pub struct I16x1(i16);
-impl_simd_base_overloads!(I16x1);
+define_simd_type!(i16, 1, i16);
 impl_simd_int_overloads!(I16x1);
 
-impl SimdBase for I16x1 {
-    const WIDTH: usize = 1;
-    type Scalar = i16;
-    type ArrayRepresentation = [i16; 1];
-    type UnderlyingType = i16;
-
+impl InternalSimdBaseIo for I16x1 {
     #[inline(always)]
     unsafe fn zeroes() -> Self {
         I16x1(0)
@@ -30,6 +22,48 @@ impl SimdBase for I16x1 {
         I16x1(x)
     }
 
+    #[inline(always)]
+    unsafe fn load_from_array(array: Self::ArrayRepresentation) -> Self {
+        I16x1(array[0])
+    }
+
+    #[inline(always)]
+    unsafe fn load_from_ptr_unaligned(ptr: *const Self::Scalar) -> Self {
+        I16x1(*ptr)
+    }
+
+    #[inline(always)]
+    unsafe fn copy_to_ptr_unaligned(self, ptr: *mut Self::Scalar) {
+        *ptr = self.0;
+    }
+
+    #[inline(always)]
+    unsafe fn load_from_ptr_aligned(ptr: *const Self::Scalar) -> Self {
+        I16x1(*ptr)
+    }
+
+    #[inline(always)]
+    unsafe fn copy_to_ptr_aligned(self, ptr: *mut Self::Scalar) {
+        *ptr = self.0;
+    }
+
+    #[inline(always)]
+    unsafe fn underlying_value(self) -> Self::UnderlyingType {
+        self.0
+    }
+
+    #[inline(always)]
+    unsafe fn underlying_value_mut(&mut self) -> &mut Self::UnderlyingType {
+        &mut self.0
+    }
+
+    #[inline(always)]
+    unsafe fn from_underlying_value(value: Self::UnderlyingType) -> Self {
+        I16x1(value)
+    }
+}
+
+impl SimdBaseOps for I16x1 {
     #[inline(always)]
     fn add(self, rhs: Self) -> Self {
         I16x1(self.0.wrapping_add(rhs.0))
@@ -119,15 +153,44 @@ impl SimdBase for I16x1 {
     fn min(self, rhs: Self) -> Self {
         I16x1(self.0.min(rhs.0))
     }
+}
+
+impl SimdInt for I16x1 {
+    #[inline(always)]
+    fn shl(self, rhs: i32) -> Self {
+        I16x1(((self.0 as u16) << rhs) as i16)
+    }
+
+    #[inline(always)]
+    fn shr(self, rhs: i32) -> Self {
+        I16x1(((self.0 as u16) >> rhs) as i16)
+    }
+}
+
+impl SimdInt16 for I16x1 {}
+
+define_simd_type!(i32, 1, i32);
+impl_simd_int_overloads!(I32x1);
+
+impl InternalSimdBaseIo for I32x1 {
+    #[inline(always)]
+    unsafe fn zeroes() -> Self {
+        I32x1(0)
+    }
+
+    #[inline(always)]
+    unsafe fn set1(x: Self::Scalar) -> Self {
+        I32x1(x)
+    }
 
     #[inline(always)]
     unsafe fn load_from_array(array: Self::ArrayRepresentation) -> Self {
-        I16x1(array[0])
+        I32x1(array[0])
     }
 
     #[inline(always)]
     unsafe fn load_from_ptr_unaligned(ptr: *const Self::Scalar) -> Self {
-        I16x1(*ptr)
+        I32x1(*ptr)
     }
 
     #[inline(always)]
@@ -137,7 +200,7 @@ impl SimdBase for I16x1 {
 
     #[inline(always)]
     unsafe fn load_from_ptr_aligned(ptr: *const Self::Scalar) -> Self {
-        I16x1(*ptr)
+        I32x1(*ptr)
     }
 
     #[inline(always)]
@@ -157,48 +220,11 @@ impl SimdBase for I16x1 {
 
     #[inline(always)]
     unsafe fn from_underlying_value(value: Self::UnderlyingType) -> Self {
-        I16x1(value)
+        I32x1(value)
     }
 }
 
-impl SimdInt for I16x1 {
-    #[inline(always)]
-    fn shl(self, rhs: i32) -> Self {
-        I16x1(((self.0 as u16) << rhs) as i16)
-    }
-
-    #[inline(always)]
-    fn shr(self, rhs: i32) -> Self {
-        I16x1(((self.0 as u16) >> rhs) as i16)
-    }
-}
-
-impl SimdInt16 for I16x1 {}
-
-// Newtypes for i32 vectors
-// We have to do this to allow for overloading of
-// __m128i etc
-#[derive(Copy, Clone)]
-pub struct I32x1(i32);
-impl_simd_base_overloads!(I32x1);
-impl_simd_int_overloads!(I32x1);
-
-impl SimdBase for I32x1 {
-    const WIDTH: usize = 1;
-    type Scalar = i32;
-    type ArrayRepresentation = [i32; 1];
-    type UnderlyingType = i32;
-
-    #[inline(always)]
-    unsafe fn zeroes() -> Self {
-        I32x1(0)
-    }
-
-    #[inline(always)]
-    unsafe fn set1(x: Self::Scalar) -> Self {
-        I32x1(x)
-    }
-
+impl SimdBaseOps for I32x1 {
     #[inline(always)]
     fn add(self, rhs: Self) -> Self {
         I32x1(self.0.wrapping_add(rhs.0))
@@ -288,46 +314,6 @@ impl SimdBase for I32x1 {
     fn min(self, rhs: Self) -> Self {
         I32x1(self.0.min(rhs.0))
     }
-
-    #[inline(always)]
-    unsafe fn load_from_array(array: Self::ArrayRepresentation) -> Self {
-        I32x1(array[0])
-    }
-
-    #[inline(always)]
-    unsafe fn load_from_ptr_unaligned(ptr: *const Self::Scalar) -> Self {
-        I32x1(*ptr)
-    }
-
-    #[inline(always)]
-    unsafe fn copy_to_ptr_unaligned(self, ptr: *mut Self::Scalar) {
-        *ptr = self.0;
-    }
-
-    #[inline(always)]
-    unsafe fn load_from_ptr_aligned(ptr: *const Self::Scalar) -> Self {
-        I32x1(*ptr)
-    }
-
-    #[inline(always)]
-    unsafe fn copy_to_ptr_aligned(self, ptr: *mut Self::Scalar) {
-        *ptr = self.0;
-    }
-
-    #[inline(always)]
-    unsafe fn underlying_value(self) -> Self::UnderlyingType {
-        self.0
-    }
-
-    #[inline(always)]
-    unsafe fn underlying_value_mut(&mut self) -> &mut Self::UnderlyingType {
-        &mut self.0
-    }
-
-    #[inline(always)]
-    unsafe fn from_underlying_value(value: Self::UnderlyingType) -> Self {
-        I32x1(value)
-    }
 }
 
 impl SimdInt for I32x1 {
@@ -356,20 +342,10 @@ impl SimdInt32 for I32x1 {
     }
 }
 
-// Newtypes for i64 vectors
-// We have to do this to allow for overloading of
-// __m128i etc
-#[derive(Copy, Clone)]
-pub struct I64x1(i64);
-impl_simd_base_overloads!(I64x1);
+define_simd_type!(i64, 1, i64);
 impl_simd_int_overloads!(I64x1);
 
-impl SimdBase for I64x1 {
-    const WIDTH: usize = 1;
-    type Scalar = i64;
-    type ArrayRepresentation = [i64; 1];
-    type UnderlyingType = i64;
-
+impl InternalSimdBaseIo for I64x1 {
     #[inline(always)]
     unsafe fn zeroes() -> Self {
         I64x1(0)
@@ -380,6 +356,48 @@ impl SimdBase for I64x1 {
         I64x1(x)
     }
 
+    #[inline(always)]
+    unsafe fn load_from_array(array: Self::ArrayRepresentation) -> Self {
+        I64x1(array[0])
+    }
+
+    #[inline(always)]
+    unsafe fn load_from_ptr_unaligned(ptr: *const Self::Scalar) -> Self {
+        I64x1(*ptr)
+    }
+
+    #[inline(always)]
+    unsafe fn copy_to_ptr_unaligned(self, ptr: *mut Self::Scalar) {
+        *ptr = self.0;
+    }
+
+    #[inline(always)]
+    unsafe fn load_from_ptr_aligned(ptr: *const Self::Scalar) -> Self {
+        I64x1(*ptr)
+    }
+
+    #[inline(always)]
+    unsafe fn copy_to_ptr_aligned(self, ptr: *mut Self::Scalar) {
+        *ptr = self.0;
+    }
+
+    #[inline(always)]
+    unsafe fn underlying_value(self) -> Self::UnderlyingType {
+        self.0
+    }
+
+    #[inline(always)]
+    unsafe fn underlying_value_mut(&mut self) -> &mut Self::UnderlyingType {
+        &mut self.0
+    }
+
+    #[inline(always)]
+    unsafe fn from_underlying_value(value: Self::UnderlyingType) -> Self {
+        I64x1(value)
+    }
+}
+
+impl SimdBaseOps for I64x1 {
     #[inline(always)]
     fn add(self, rhs: Self) -> Self {
         I64x1(self.0.wrapping_add(rhs.0))
@@ -469,46 +487,6 @@ impl SimdBase for I64x1 {
     fn min(self, rhs: Self) -> Self {
         I64x1(self.0.min(rhs.0))
     }
-
-    #[inline(always)]
-    unsafe fn load_from_array(array: Self::ArrayRepresentation) -> Self {
-        I64x1(array[0])
-    }
-
-    #[inline(always)]
-    unsafe fn load_from_ptr_unaligned(ptr: *const Self::Scalar) -> Self {
-        I64x1(*ptr)
-    }
-
-    #[inline(always)]
-    unsafe fn copy_to_ptr_unaligned(self, ptr: *mut Self::Scalar) {
-        *ptr = self.0;
-    }
-
-    #[inline(always)]
-    unsafe fn load_from_ptr_aligned(ptr: *const Self::Scalar) -> Self {
-        I64x1(*ptr)
-    }
-
-    #[inline(always)]
-    unsafe fn copy_to_ptr_aligned(self, ptr: *mut Self::Scalar) {
-        *ptr = self.0;
-    }
-
-    #[inline(always)]
-    unsafe fn underlying_value(self) -> Self::UnderlyingType {
-        self.0
-    }
-
-    #[inline(always)]
-    unsafe fn underlying_value_mut(&mut self) -> &mut Self::UnderlyingType {
-        &mut self.0
-    }
-
-    #[inline(always)]
-    unsafe fn from_underlying_value(value: Self::UnderlyingType) -> Self {
-        I64x1(value)
-    }
 }
 
 impl SimdInt for I64x1 {
@@ -537,20 +515,10 @@ impl SimdInt64 for I64x1 {
     }
 }
 
-// Newtypes for f32 vectors
-// We have to do this to allow for overloading of
-// __m128 etc
-#[derive(Copy, Clone)]
-pub struct F32x1(f32);
-impl_simd_base_overloads!(F32x1);
+define_simd_type!(f32, 1, f32);
 impl_simd_float_overloads!(F32x1);
 
-impl SimdBase for F32x1 {
-    const WIDTH: usize = 1;
-    type Scalar = f32;
-    type ArrayRepresentation = [f32; 1];
-    type UnderlyingType = f32;
-
+impl InternalSimdBaseIo for F32x1 {
     #[inline(always)]
     unsafe fn zeroes() -> Self {
         F32x1(0.0)
@@ -561,6 +529,48 @@ impl SimdBase for F32x1 {
         F32x1(x)
     }
 
+    #[inline(always)]
+    unsafe fn load_from_array(array: Self::ArrayRepresentation) -> Self {
+        F32x1(array[0])
+    }
+
+    #[inline(always)]
+    unsafe fn load_from_ptr_unaligned(ptr: *const Self::Scalar) -> Self {
+        F32x1(*ptr)
+    }
+
+    #[inline(always)]
+    unsafe fn copy_to_ptr_unaligned(self, ptr: *mut Self::Scalar) {
+        *ptr = self.0;
+    }
+
+    #[inline(always)]
+    unsafe fn load_from_ptr_aligned(ptr: *const Self::Scalar) -> Self {
+        F32x1(*ptr)
+    }
+
+    #[inline(always)]
+    unsafe fn copy_to_ptr_aligned(self, ptr: *mut Self::Scalar) {
+        *ptr = self.0;
+    }
+
+    #[inline(always)]
+    unsafe fn underlying_value(self) -> Self::UnderlyingType {
+        self.0
+    }
+
+    #[inline(always)]
+    unsafe fn underlying_value_mut(&mut self) -> &mut Self::UnderlyingType {
+        &mut self.0
+    }
+
+    #[inline(always)]
+    unsafe fn from_underlying_value(value: Self::UnderlyingType) -> Self {
+        F32x1(value)
+    }
+}
+
+impl SimdBaseOps for F32x1 {
     #[inline(always)]
     fn add(self, rhs: Self) -> Self {
         F32x1(self.0 + rhs.0)
@@ -674,46 +684,6 @@ impl SimdBase for F32x1 {
     fn min(self, rhs: Self) -> Self {
         F32x1(self.0.min(rhs.0))
     }
-
-    #[inline(always)]
-    unsafe fn load_from_array(array: Self::ArrayRepresentation) -> Self {
-        F32x1(array[0])
-    }
-
-    #[inline(always)]
-    unsafe fn load_from_ptr_unaligned(ptr: *const Self::Scalar) -> Self {
-        F32x1(*ptr)
-    }
-
-    #[inline(always)]
-    unsafe fn copy_to_ptr_unaligned(self, ptr: *mut Self::Scalar) {
-        *ptr = self.0;
-    }
-
-    #[inline(always)]
-    unsafe fn load_from_ptr_aligned(ptr: *const Self::Scalar) -> Self {
-        F32x1(*ptr)
-    }
-
-    #[inline(always)]
-    unsafe fn copy_to_ptr_aligned(self, ptr: *mut Self::Scalar) {
-        *ptr = self.0;
-    }
-
-    #[inline(always)]
-    unsafe fn underlying_value(self) -> Self::UnderlyingType {
-        self.0
-    }
-
-    #[inline(always)]
-    unsafe fn underlying_value_mut(&mut self) -> &mut Self::UnderlyingType {
-        &mut self.0
-    }
-
-    #[inline(always)]
-    unsafe fn from_underlying_value(value: Self::UnderlyingType) -> Self {
-        F32x1(value)
-    }
 }
 
 impl SimdFloat for F32x1 {
@@ -807,18 +777,10 @@ impl SimdFloat32 for F32x1 {
     }
 }
 
-// Newtypes for f64 vectors
-#[derive(Copy, Clone)]
-pub struct F64x1(f64);
-impl_simd_base_overloads!(F64x1);
+define_simd_type!(f64, 1, f64);
 impl_simd_float_overloads!(F64x1);
 
-impl SimdBase for F64x1 {
-    const WIDTH: usize = 1;
-    type Scalar = f64;
-    type ArrayRepresentation = [f64; 1];
-    type UnderlyingType = f64;
-
+impl InternalSimdBaseIo for F64x1 {
     #[inline(always)]
     unsafe fn zeroes() -> Self {
         F64x1(0.0)
@@ -829,6 +791,48 @@ impl SimdBase for F64x1 {
         F64x1(x)
     }
 
+    #[inline(always)]
+    unsafe fn load_from_array(array: Self::ArrayRepresentation) -> Self {
+        F64x1(array[0])
+    }
+
+    #[inline(always)]
+    unsafe fn load_from_ptr_unaligned(ptr: *const Self::Scalar) -> Self {
+        F64x1(*ptr)
+    }
+
+    #[inline(always)]
+    unsafe fn copy_to_ptr_unaligned(self, ptr: *mut Self::Scalar) {
+        *ptr = self.0;
+    }
+
+    #[inline(always)]
+    unsafe fn load_from_ptr_aligned(ptr: *const Self::Scalar) -> Self {
+        F64x1(*ptr)
+    }
+
+    #[inline(always)]
+    unsafe fn copy_to_ptr_aligned(self, ptr: *mut Self::Scalar) {
+        *ptr = self.0;
+    }
+
+    #[inline(always)]
+    unsafe fn underlying_value(self) -> Self::UnderlyingType {
+        self.0
+    }
+
+    #[inline(always)]
+    unsafe fn underlying_value_mut(&mut self) -> &mut Self::UnderlyingType {
+        &mut self.0
+    }
+
+    #[inline(always)]
+    unsafe fn from_underlying_value(value: Self::UnderlyingType) -> Self {
+        F64x1(value)
+    }
+}
+
+impl SimdBaseOps for F64x1 {
     #[inline(always)]
     fn add(self, rhs: Self) -> Self {
         F64x1(self.0 + rhs.0)
@@ -941,46 +945,6 @@ impl SimdBase for F64x1 {
     #[inline(always)]
     fn min(self, rhs: Self) -> Self {
         F64x1(self.0.min(rhs.0))
-    }
-
-    #[inline(always)]
-    unsafe fn load_from_array(array: Self::ArrayRepresentation) -> Self {
-        F64x1(array[0])
-    }
-
-    #[inline(always)]
-    unsafe fn load_from_ptr_unaligned(ptr: *const Self::Scalar) -> Self {
-        F64x1(*ptr)
-    }
-
-    #[inline(always)]
-    unsafe fn copy_to_ptr_unaligned(self, ptr: *mut Self::Scalar) {
-        *ptr = self.0;
-    }
-
-    #[inline(always)]
-    unsafe fn load_from_ptr_aligned(ptr: *const Self::Scalar) -> Self {
-        F64x1(*ptr)
-    }
-
-    #[inline(always)]
-    unsafe fn copy_to_ptr_aligned(self, ptr: *mut Self::Scalar) {
-        *ptr = self.0;
-    }
-
-    #[inline(always)]
-    unsafe fn underlying_value(self) -> Self::UnderlyingType {
-        self.0
-    }
-
-    #[inline(always)]
-    unsafe fn underlying_value_mut(&mut self) -> &mut Self::UnderlyingType {
-        &mut self.0
-    }
-
-    #[inline(always)]
-    unsafe fn from_underlying_value(value: Self::UnderlyingType) -> Self {
-        F64x1(value)
     }
 }
