+++ src/engines/sse41/mod.rs
@@ -1,65 +1,51 @@
-use crate::libm_ext::FloatExt;
+use crate::{
+    libm_ext::FloatExt, InternalSimdBaseIo, SimdBaseOps, SimdConsts, SimdFloat, SimdFloat32,
+    SimdFloat64, SimdInt, SimdInt16, SimdInt32, SimdInt64,
+};
 
-use super::*;
 #[cfg(target_arch = "x86")]
 use core::arch::x86::*;
 
 #[cfg(target_arch = "x86_64")]
 use core::arch::x86_64::*;
 
+use core::ops::*;
+
 mod simd;
-pub use self::overloads::*;
 pub use self::simd::*;
 
-#[derive(Copy, Clone)]
-pub struct I16x8(__m128i);
-impl_simd_base_overloads!(I16x8);
-impl_simd_int_overloads!(I16x8);
+define_simd_type!(i16, 8, __m128i, _41);
+impl_simd_int_overloads!(I16x8_41);
 
-impl SimdBase for I16x8 {
-    const WIDTH: usize = 8;
-    type Scalar = i16;
-    type ArrayRepresentation = [i16; 8];
-    type UnderlyingType = __m128i;
-
-    #[inline(always)]
-    unsafe fn zeroes() -> Self {
-        I16x8(_mm_setzero_si128())
-    }
-
-    #[inline(always)]
-    unsafe fn set1(x: Self::Scalar) -> Self {
-        I16x8(_mm_set1_epi16(x))
-    }
-
+impl SimdBaseOps for I16x8_41 {
     #[inline(always)]
     fn add(self, rhs: Self) -> Self {
-        unsafe { I16x8(_mm_add_epi16(self.0, rhs.0)) }
+        unsafe { I16x8_41(_mm_add_epi16(self.0, rhs.0)) }
     }
 
     #[inline(always)]
     fn sub(self, rhs: Self) -> Self {
-        unsafe { I16x8(_mm_sub_epi16(self.0, rhs.0)) }
+        unsafe { I16x8_41(_mm_sub_epi16(self.0, rhs.0)) }
     }
 
     #[inline(always)]
     fn mul(self, rhs: Self) -> Self {
-        unsafe { I16x8(_mm_mullo_epi16(self.0, rhs.0)) }
+        unsafe { I16x8_41(_mm_mullo_epi16(self.0, rhs.0)) }
     }
 
     #[inline(always)]
     fn bit_and(self, rhs: Self) -> Self {
-        unsafe { I16x8(_mm_and_si128(self.0, rhs.0)) }
+        unsafe { I16x8_41(_mm_and_si128(self.0, rhs.0)) }
     }
 
     #[inline(always)]
     fn bit_or(self, rhs: Self) -> Self {
-        unsafe { I16x8(_mm_or_si128(self.0, rhs.0)) }
+        unsafe { I16x8_41(_mm_or_si128(self.0, rhs.0)) }
     }
 
     #[inline(always)]
     fn bit_xor(self, rhs: Self) -> Self {
-        unsafe { I16x8(_mm_xor_si128(self.0, rhs.0)) }
+        unsafe { I16x8_41(_mm_xor_si128(self.0, rhs.0)) }
     }
 
     #[inline(always)]
@@ -77,7 +63,7 @@ impl SimdBase for I16x8 {
 
     #[inline(always)]
     fn and_not(self, rhs: Self) -> Self {
-        unsafe { I16x8(_mm_andnot_si128(self.0, rhs.0)) }
+        unsafe { I16x8_41(_mm_andnot_si128(self.0, rhs.0)) }
     }
 
     #[inline(always)]
@@ -92,7 +78,7 @@ impl SimdBase for I16x8 {
 
     #[inline(always)]
     fn cmp_eq(self, rhs: Self) -> Self {
-        unsafe { I16x8(_mm_cmpeq_epi16(self.0, rhs.0)) }
+        unsafe { I16x8_41(_mm_cmpeq_epi16(self.0, rhs.0)) }
     }
 
     #[inline(always)]
@@ -102,7 +88,7 @@ impl SimdBase for I16x8 {
 
     #[inline(always)]
     fn cmp_lt(self, rhs: Self) -> Self {
-        unsafe { I16x8(_mm_cmplt_epi16(self.0, rhs.0)) }
+        unsafe { I16x8_41(_mm_cmplt_epi16(self.0, rhs.0)) }
     }
 
     #[inline(always)]
@@ -112,7 +98,7 @@ impl SimdBase for I16x8 {
 
     #[inline(always)]
     fn cmp_gt(self, rhs: Self) -> Self {
-        unsafe { I16x8(_mm_cmpgt_epi16(self.0, rhs.0)) }
+        unsafe { I16x8_41(_mm_cmpgt_epi16(self.0, rhs.0)) }
     }
 
     #[inline(always)]
@@ -122,22 +108,34 @@ impl SimdBase for I16x8 {
 
     #[inline(always)]
     fn max(self, rhs: Self) -> Self {
-        unsafe { I16x8(_mm_max_epi16(self.0, rhs.0)) }
+        unsafe { I16x8_41(_mm_max_epi16(self.0, rhs.0)) }
     }
 
     #[inline(always)]
     fn min(self, rhs: Self) -> Self {
-        unsafe { I16x8(_mm_min_epi16(self.0, rhs.0)) }
+        unsafe { I16x8_41(_mm_min_epi16(self.0, rhs.0)) }
+    }
+}
+
+impl InternalSimdBaseIo for I16x8_41 {
+    #[inline(always)]
+    unsafe fn zeroes() -> Self {
+        I16x8_41(_mm_setzero_si128())
+    }
+
+    #[inline(always)]
+    unsafe fn set1(x: Self::Scalar) -> Self {
+        I16x8_41(_mm_set1_epi16(x))
     }
 
     #[inline(always)]
     unsafe fn load_from_array(array: Self::ArrayRepresentation) -> Self {
-        I16x8(_mm_loadu_si128(array.as_ptr() as *const __m128i))
+        I16x8_41(_mm_loadu_si128(array.as_ptr() as *const __m128i))
     }
 
     #[inline(always)]
     unsafe fn load_from_ptr_unaligned(ptr: *const Self::Scalar) -> Self {
-        I16x8(_mm_loadu_si128(ptr as *const __m128i))
+        I16x8_41(_mm_loadu_si128(ptr as *const __m128i))
     }
 
     #[inline(always)]
@@ -147,7 +145,7 @@ impl SimdBase for I16x8 {
 
     #[inline(always)]
     unsafe fn load_from_ptr_aligned(ptr: *const Self::Scalar) -> Self {
-        I16x8(_mm_load_si128(ptr as *const __m128i))
+        I16x8_41(_mm_load_si128(ptr as *const __m128i))
     }
 
     #[inline(always)]
@@ -167,16 +165,16 @@ impl SimdBase for I16x8 {
 
     #[inline(always)]
     unsafe fn from_underlying_value(value: Self::UnderlyingType) -> Self {
-        I16x8(value)
+        I16x8_41(value)
     }
 }
 
-impl SimdInt for I16x8 {
+impl SimdInt for I16x8_41 {
     #[inline(always)]
     fn shl(self, rhs: i32) -> Self {
         unsafe {
             let rhs = _mm_cvtsi32_si128(rhs);
-            I16x8(_mm_sll_epi16(self.0, rhs))
+            I16x8_41(_mm_sll_epi16(self.0, rhs))
         }
     }
 
@@ -184,79 +182,55 @@ impl SimdInt for I16x8 {
     fn shr(self, rhs: i32) -> Self {
         unsafe {
             let rhs = _mm_cvtsi32_si128(rhs);
-            I16x8(_mm_srl_epi16(self.0, rhs))
+            I16x8_41(_mm_srl_epi16(self.0, rhs))
         }
     }
 
     #[inline(always)]
     fn shl_const<const BY: i32>(self) -> Self {
-        unsafe { I16x8(_mm_slli_epi16(self.0, BY)) }
+        unsafe { I16x8_41(_mm_slli_epi16(self.0, BY)) }
     }
 
     #[inline(always)]
     fn shr_const<const BY: i32>(self) -> Self {
-        unsafe { I16x8(_mm_srli_epi16(self.0, BY)) }
+        unsafe { I16x8_41(_mm_srli_epi16(self.0, BY)) }
     }
 }
 
-impl SimdInt16 for I16x8 {}
+impl SimdInt16 for I16x8_41 {}
 
-#[derive(Copy, Clone)]
-pub struct I32x4(__m128i);
-impl_simd_base_overloads!(I32x4);
-impl_simd_int_overloads!(I32x4);
+define_simd_type!(i32, 4, __m128i, _41);
+impl_simd_int_overloads!(I32x4_41);
 
-impl SimdBase for I32x4 {
-    const WIDTH: usize = 4;
-    type Scalar = i32;
-    type ArrayRepresentation = [i32; 4];
-    type UnderlyingType = __m128i;
-
-    #[inline(always)]
-    unsafe fn zeroes() -> Self {
-        I32x4(_mm_setzero_si128())
-    }
-
-    #[inline(always)]
-    unsafe fn set1(x: Self::Scalar) -> Self {
-        I32x4(_mm_set1_epi32(x))
-    }
-
+impl SimdBaseOps for I32x4_41 {
     #[inline(always)]
     fn add(self, rhs: Self) -> Self {
-        unsafe { I32x4(_mm_add_epi32(self.0, rhs.0)) }
+        unsafe { I32x4_41(_mm_add_epi32(self.0, rhs.0)) }
     }
 
     #[inline(always)]
     fn sub(self, rhs: Self) -> Self {
-        unsafe { I32x4(_mm_sub_epi32(self.0, rhs.0)) }
+        unsafe { I32x4_41(_mm_sub_epi32(self.0, rhs.0)) }
     }
 
     #[inline(always)]
     fn mul(self, rhs: Self) -> Self {
-        unsafe {
-            Self::load_from_array([
-                self[0].wrapping_mul(rhs[0]),
-                self[1].wrapping_mul(rhs[1]),
-                self[2].wrapping_mul(rhs[2]),
-                self[3].wrapping_mul(rhs[3]),
-            ])
-        }
+        unsafe { I32x4_41(_mm_mullo_epi32(self.0, rhs.0)) }
     }
 
     #[inline(always)]
     fn bit_and(self, rhs: Self) -> Self {
-        unsafe { I32x4(_mm_and_si128(self.0, rhs.0)) }
+        unsafe { I32x4_41(_mm_and_si128(self.0, rhs.0)) }
     }
 
     #[inline(always)]
     fn bit_or(self, rhs: Self) -> Self {
-        unsafe { I32x4(_mm_or_si128(self.0, rhs.0)) }
+        unsafe { I32x4_41(_mm_or_si128(self.0, rhs.0)) }
     }
 
     #[inline(always)]
     fn bit_xor(self, rhs: Self) -> Self {
-        unsafe { I32x4(_mm_xor_si128(self.0, rhs.0)) }
+        unsafe { I32x4_41(_mm_xor_si128(self.0, rhs.0)) }
     }
 
     #[inline(always)]
@@ -274,22 +248,19 @@ impl SimdBase for I32x4 {
 
     #[inline(always)]
     fn and_not(self, rhs: Self) -> Self {
-        unsafe { I32x4(_mm_andnot_si128(self.0, rhs.0)) }
+        unsafe { I32x4_41(_mm_andnot_si128(self.0, rhs.0)) }
     }
 
     #[inline(always)]
     fn blendv(self, a: Self, b: Self) -> Self {
-        unsafe {
-            let mask = self.cmp_eq(Self::set1(-1));
-            let a = mask.and_not(a);
-            let b = mask.bit_and(b);
-            a.bit_or(b)
-        }
+        self.bitcast_f32()
+            .blendv(a.bitcast_f32(), b.bitcast_f32())
+            .bitcast_i32()
     }
 
     #[inline(always)]
     fn cmp_eq(self, rhs: Self) -> Self {
-        unsafe { I32x4(_mm_cmpeq_epi32(self.0, rhs.0)) }
+        unsafe { I32x4_41(_mm_cmpeq_epi32(self.0, rhs.0)) }
     }
 
     #[inline(always)]
@@ -299,7 +270,7 @@ impl SimdBase for I32x4 {
 
     #[inline(always)]
     fn cmp_lt(self, rhs: Self) -> Self {
-        unsafe { I32x4(_mm_cmplt_epi32(self.0, rhs.0)) }
+        unsafe { I32x4_41(_mm_cmplt_epi32(self.0, rhs.0)) }
     }
 
     #[inline(always)]
@@ -309,7 +280,7 @@ impl SimdBase for I32x4 {
 
     #[inline(always)]
     fn cmp_gt(self, rhs: Self) -> Self {
-        unsafe { I32x4(_mm_cmpgt_epi32(self.0, rhs.0)) }
+        unsafe { I32x4_41(_mm_cmpgt_epi32(self.0, rhs.0)) }
     }
 
     #[inline(always)]
@@ -319,24 +290,34 @@ impl SimdBase for I32x4 {
 
     #[inline(always)]
     fn max(self, rhs: Self) -> Self {
-        let mask = self.cmp_gt(rhs);
-        mask.blendv(rhs, self)
+        unsafe { I32x4_41(_mm_max_epi32(self.0, rhs.0)) }
     }
 
     #[inline(always)]
     fn min(self, rhs: Self) -> Self {
-        let mask = self.cmp_gt(rhs);
-        mask.blendv(self, rhs)
+        unsafe { I32x4_41(_mm_min_epi32(self.0, rhs.0)) }
+    }
+}
+
+impl InternalSimdBaseIo for I32x4_41 {
+    #[inline(always)]
+    unsafe fn zeroes() -> Self {
+        I32x4_41(_mm_setzero_si128())
+    }
+
+    #[inline(always)]
+    unsafe fn set1(x: Self::Scalar) -> Self {
+        I32x4_41(_mm_set1_epi32(x))
     }
 
     #[inline(always)]
     unsafe fn load_from_array(array: Self::ArrayRepresentation) -> Self {
-        I32x4(_mm_loadu_si128(array.as_ptr() as *const __m128i))
+        I32x4_41(_mm_loadu_si128(array.as_ptr() as *const __m128i))
     }
 
     #[inline(always)]
     unsafe fn load_from_ptr_unaligned(ptr: *const Self::Scalar) -> Self {
-        I32x4(_mm_loadu_si128(ptr as *const __m128i))
+        I32x4_41(_mm_loadu_si128(ptr as *const __m128i))
     }
 
     #[inline(always)]
@@ -346,7 +327,7 @@ impl SimdBase for I32x4 {
 
     #[inline(always)]
     unsafe fn load_from_ptr_aligned(ptr: *const Self::Scalar) -> Self {
-        I32x4(_mm_load_si128(ptr as *const __m128i))
+        I32x4_41(_mm_load_si128(ptr as *const __m128i))
     }
 
     #[inline(always)]
@@ -366,16 +347,16 @@ impl SimdBase for I32x4 {
 
     #[inline(always)]
     unsafe fn from_underlying_value(value: Self::UnderlyingType) -> Self {
-        I32x4(value)
+        I32x4_41(value)
     }
 }
 
-impl SimdInt for I32x4 {
+impl SimdInt for I32x4_41 {
     #[inline(always)]
     fn shl(self, rhs: i32) -> Self {
         unsafe {
             let rhs = _mm_cvtsi32_si128(rhs);
-            I32x4(_mm_sll_epi32(self.0, rhs))
+            I32x4_41(_mm_sll_epi32(self.0, rhs))
         }
     }
 
@@ -383,64 +364,47 @@ impl SimdInt for I32x4 {
     fn shr(self, rhs: i32) -> Self {
         unsafe {
             let rhs = _mm_cvtsi32_si128(rhs);
-            I32x4(_mm_srl_epi32(self.0, rhs))
+            I32x4_41(_mm_srl_epi32(self.0, rhs))
         }
     }
 
     #[inline(always)]
     fn shl_const<const BY: i32>(self) -> Self {
-        unsafe { I32x4(_mm_slli_epi32(self.0, BY)) }
+        unsafe { I32x4_41(_mm_slli_epi32(self.0, BY)) }
     }
 
     #[inline(always)]
     fn shr_const<const BY: i32>(self) -> Self {
-        unsafe { I32x4(_mm_srli_epi32(self.0, BY)) }
+        unsafe { I32x4_41(_mm_srli_epi32(self.0, BY)) }
     }
 }
 
-impl SimdInt32 for I32x4 {
-    type SimdF32 = F32x4;
+impl SimdInt32 for I32x4_41 {
+    type SimdF32 = F32x4_41;
 
     #[inline(always)]
     fn bitcast_f32(self) -> Self::SimdF32 {
-        unsafe { F32x4(_mm_castsi128_ps(self.0)) }
+        unsafe { F32x4_41(_mm_castsi128_ps(self.0)) }
     }
 
     #[inline(always)]
     fn cast_f32(self) -> Self::SimdF32 {
-        unsafe { F32x4(_mm_cvtepi32_ps(self.0)) }
+        unsafe { F32x4_41(_mm_cvtepi32_ps(self.0)) }
     }
 }
 
-#[derive(Copy, Clone)]
-pub struct I64x2(__m128i);
-impl_simd_base_overloads!(I64x2);
-impl_simd_int_overloads!(I64x2);
+define_simd_type!(i64, 2, __m128i, _41);
+impl_simd_int_overloads!(I64x2_41);
 
-impl SimdBase for I64x2 {
-    const WIDTH: usize = 2;
-    type Scalar = i64;
-    type ArrayRepresentation = [i64; 2];
-    type UnderlyingType = __m128i;
-
-    #[inline(always)]
-    unsafe fn zeroes() -> Self {
-        I64x2(_mm_setzero_si128())
-    }
-
-    #[inline(always)]
-    unsafe fn set1(x: Self::Scalar) -> Self {
-        I64x2(_mm_set1_epi64x(x))
-    }
-
+impl SimdBaseOps for I64x2_41 {
     #[inline(always)]
     fn add(self, rhs: Self) -> Self {
-        unsafe { I64x2(_mm_add_epi64(self.0, rhs.0)) }
+        unsafe { I64x2_41(_mm_add_epi64(self.0, rhs.0)) }
     }
 
     #[inline(always)]
     fn sub(self, rhs: Self) -> Self {
-        unsafe { I64x2(_mm_sub_epi64(self.0, rhs.0)) }
+        unsafe { I64x2_41(_mm_sub_epi64(self.0, rhs.0)) }
     }
 
     #[inline(always)]
@@ -452,17 +416,17 @@ impl SimdBase for I64x2 {
 
     #[inline(always)]
     fn bit_and(self, rhs: Self) -> Self {
-        unsafe { I64x2(_mm_and_si128(self.0, rhs.0)) }
+        unsafe { I64x2_41(_mm_and_si128(self.0, rhs.0)) }
     }
 
     #[inline(always)]
     fn bit_or(self, rhs: Self) -> Self {
-        unsafe { I64x2(_mm_or_si128(self.0, rhs.0)) }
+        unsafe { I64x2_41(_mm_or_si128(self.0, rhs.0)) }
     }
 
     #[inline(always)]
     fn bit_xor(self, rhs: Self) -> Self {
-        unsafe { I64x2(_mm_xor_si128(self.0, rhs.0)) }
+        unsafe { I64x2_41(_mm_xor_si128(self.0, rhs.0)) }
     }
 
     #[inline(always)]
@@ -480,27 +444,19 @@ impl SimdBase for I64x2 {
 
     #[inline(always)]
     fn and_not(self, rhs: Self) -> Self {
-        unsafe { I64x2(_mm_andnot_si128(self.0, rhs.0)) }
+        unsafe { I64x2_41(_mm_andnot_si128(self.0, rhs.0)) }
     }
 
     #[inline(always)]
     fn blendv(self, a: Self, b: Self) -> Self {
-        unsafe {
-            let mask = self.cmp_eq(Self::set1(-1));
-            let a = mask.and_not(a);
-            let b = mask.bit_and(b);
-            a.bit_or(b)
-        }
+        self.bitcast_f64()
+            .blendv(a.bitcast_f64(), b.bitcast_f64())
+            .bitcast_i64()
     }
 
     #[inline(always)]
     fn cmp_eq(self, rhs: Self) -> Self {
-        unsafe {
-            let com32 = _mm_cmpeq_epi32(self.0, rhs.0); // 32 bit compares
-            let com32s = _mm_shuffle_epi32(com32, 0xB1); // swap low and high dwords
-            let test = _mm_and_si128(com32, com32s); // low & high
-            I64x2(test)
-        }
+        unsafe { I64x2_41(_mm_cmpeq_epi64(self.0, rhs.0)) }
     }
 
     #[inline(always)]
@@ -556,18 +512,30 @@ impl SimdBase for I64x2 {
 
     #[inline(always)]
     fn min(self, rhs: Self) -> Self {
-        let cmp = self.cmp_gt(rhs);
-        cmp.blendv(self, rhs)
+        let cmp = self.cmp_lt(rhs);
+        cmp.blendv(rhs, self)
+    }
+}
+
+impl InternalSimdBaseIo for I64x2_41 {
+    #[inline(always)]
+    unsafe fn zeroes() -> Self {
+        I64x2_41(_mm_setzero_si128())
+    }
+
+    #[inline(always)]
+    unsafe fn set1(x: Self::Scalar) -> Self {
+        I64x2_41(_mm_set1_epi64x(x))
     }
 
     #[inline(always)]
     unsafe fn load_from_array(array: Self::ArrayRepresentation) -> Self {
-        I64x2(_mm_loadu_si128(array.as_ptr() as *const _))
+        I64x2_41(_mm_loadu_si128(array.as_ptr() as *const _))
     }
 
     #[inline(always)]
     unsafe fn load_from_ptr_unaligned(ptr: *const Self::Scalar) -> Self {
-        I64x2(_mm_loadu_si128(ptr as *const _))
+        I64x2_41(_mm_loadu_si128(ptr as *const _))
     }
 
     #[inline(always)]
@@ -577,7 +545,7 @@ impl SimdBase for I64x2 {
 
     #[inline(always)]
     unsafe fn load_from_ptr_aligned(ptr: *const Self::Scalar) -> Self {
-        I64x2(_mm_load_si128(ptr as *const _))
+        I64x2_41(_mm_load_si128(ptr as *const _))
     }
 
     #[inline(always)]
@@ -597,16 +565,16 @@ impl SimdBase for I64x2 {
 
     #[inline(always)]
     unsafe fn from_underlying_value(value: Self::UnderlyingType) -> Self {
-        I64x2(value)
+        I64x2_41(value)
     }
 }
 
-impl SimdInt for I64x2 {
+impl SimdInt for I64x2_41 {
     #[inline(always)]
     fn shl(self, rhs: i32) -> Self {
         unsafe {
             let rhs = _mm_cvtsi32_si128(rhs);
-            I64x2(_mm_sll_epi64(self.0, rhs))
+            I64x2_41(_mm_sll_epi64(self.0, rhs))
         }
     }
 
@@ -614,27 +582,27 @@ impl SimdInt for I64x2 {
     fn shr(self, rhs: i32) -> Self {
         unsafe {
             let rhs = _mm_cvtsi32_si128(rhs);
-            I64x2(_mm_srl_epi64(self.0, rhs))
+            I64x2_41(_mm_srl_epi64(self.0, rhs))
         }
     }
 
     #[inline(always)]
     fn shl_const<const BY: i32>(self) -> Self {
-        unsafe { I64x2(_mm_slli_epi64(self.0, BY)) }
+        unsafe { I64x2_41(_mm_slli_epi64(self.0, BY)) }
     }
 
     #[inline(always)]
     fn shr_const<const BY: i32>(self) -> Self {
-        unsafe { I64x2(_mm_srli_epi64(self.0, BY)) }
+        unsafe { I64x2_41(_mm_srli_epi64(self.0, BY)) }
     }
 }
 
-impl SimdInt64 for I64x2 {
-    type SimdF64 = F64x2;
+impl SimdInt64 for I64x2_41 {
+    type SimdF64 = F64x2_41;
 
     #[inline(always)]
     fn bitcast_f64(self) -> Self::SimdF64 {
-        unsafe { F64x2(_mm_castsi128_pd(self.0)) }
+        unsafe { F64x2_41(_mm_castsi128_pd(self.0)) }
     }
 
     #[inline(always)]
@@ -643,60 +611,43 @@ impl SimdInt64 for I64x2 {
     }
 }
 
-#[derive(Copy, Clone)]
-pub struct F32x4(__m128);
-impl_simd_base_overloads!(F32x4);
-impl_simd_float_overloads!(F32x4);
+define_simd_type!(f32, 4, __m128, _41);
+impl_simd_float_overloads!(F32x4_41);
 
-impl SimdBase for F32x4 {
-    const WIDTH: usize = 4;
-    type Scalar = f32;
-    type ArrayRepresentation = [f32; 4];
-    type UnderlyingType = __m128;
-
-    #[inline(always)]
-    unsafe fn zeroes() -> Self {
-        F32x4(_mm_setzero_ps())
-    }
-
-    #[inline(always)]
-    unsafe fn set1(x: Self::Scalar) -> Self {
-        F32x4(_mm_set1_ps(x))
-    }
-
+impl SimdBaseOps for F32x4_41 {
     #[inline(always)]
     fn add(self, rhs: Self) -> Self {
-        unsafe { F32x4(_mm_add_ps(self.0, rhs.0)) }
+        unsafe { F32x4_41(_mm_add_ps(self.0, rhs.0)) }
     }
 
     #[inline(always)]
     fn sub(self, rhs: Self) -> Self {
-        unsafe { F32x4(_mm_sub_ps(self.0, rhs.0)) }
+        unsafe { F32x4_41(_mm_sub_ps(self.0, rhs.0)) }
     }
 
     #[inline(always)]
     fn mul(self, rhs: Self) -> Self {
-        unsafe { F32x4(_mm_mul_ps(self.0, rhs.0)) }
+        unsafe { F32x4_41(_mm_mul_ps(self.0, rhs.0)) }
     }
 
     #[inline(always)]
     fn bit_and(self, rhs: Self) -> Self {
-        unsafe { F32x4(_mm_and_ps(self.0, rhs.0)) }
+        unsafe { F32x4_41(_mm_and_ps(self.0, rhs.0)) }
     }
 
     #[inline(always)]
     fn bit_or(self, rhs: Self) -> Self {
-        unsafe { F32x4(_mm_or_ps(self.0, rhs.0)) }
+        unsafe { F32x4_41(_mm_or_ps(self.0, rhs.0)) }
     }
 
     #[inline(always)]
     fn bit_xor(self, rhs: Self) -> Self {
-        unsafe { F32x4(_mm_xor_ps(self.0, rhs.0)) }
+        unsafe { F32x4_41(_mm_xor_ps(self.0, rhs.0)) }
     }
 
     #[inline(always)]
     fn bit_not(self) -> Self {
-        unsafe { I32x4::set1(-1).bitcast_f32().bit_xor(self) }
+        unsafe { I32x4_41::set1(-1).bitcast_f32().bit_xor(self) }
     }
 
     #[inline(always)]
@@ -706,64 +657,74 @@ impl SimdBase for F32x4 {
 
     #[inline(always)]
     fn and_not(self, rhs: Self) -> Self {
-        unsafe { F32x4(_mm_andnot_ps(self.0, rhs.0)) }
+        unsafe { F32x4_41(_mm_andnot_ps(self.0, rhs.0)) }
     }
 
     #[inline(always)]
     fn blendv(self, a: Self, b: Self) -> Self {
-        self.bitcast_i32()
-            .blendv(a.bitcast_i32(), b.bitcast_i32())
-            .bitcast_f32()
+        unsafe { F32x4_41(_mm_blendv_ps(a.0, b.0, self.0)) }
     }
 
     #[inline(always)]
     fn cmp_eq(self, rhs: Self) -> Self {
-        unsafe { F32x4(_mm_cmpeq_ps(self.0, rhs.0)) }
+        unsafe { F32x4_41(_mm_cmpeq_ps(self.0, rhs.0)) }
     }
 
     #[inline(always)]
     fn cmp_neq(self, rhs: Self) -> Self {
-        unsafe { F32x4(_mm_cmpneq_ps(self.0, rhs.0)) }
+        unsafe { F32x4_41(_mm_cmpneq_ps(self.0, rhs.0)) }
     }
 
     #[inline(always)]
     fn cmp_lt(self, rhs: Self) -> Self {
-        unsafe { F32x4(_mm_cmplt_ps(self.0, rhs.0)) }
+        unsafe { F32x4_41(_mm_cmplt_ps(self.0, rhs.0)) }
     }
 
     #[inline(always)]
     fn cmp_lte(self, rhs: Self) -> Self {
-        unsafe { F32x4(_mm_cmple_ps(self.0, rhs.0)) }
+        unsafe { F32x4_41(_mm_cmple_ps(self.0, rhs.0)) }
     }
 
     #[inline(always)]
     fn cmp_gt(self, rhs: Self) -> Self {
-        unsafe { F32x4(_mm_cmpgt_ps(self.0, rhs.0)) }
+        unsafe { F32x4_41(_mm_cmpgt_ps(self.0, rhs.0)) }
     }
 
     #[inline(always)]
     fn cmp_gte(self, rhs: Self) -> Self {
-        unsafe { F32x4(_mm_cmpge_ps(self.0, rhs.0)) }
+        unsafe { F32x4_41(_mm_cmpge_ps(self.0, rhs.0)) }
     }
 
     #[inline(always)]
     fn max(self, rhs: Self) -> Self {
-        unsafe { F32x4(_mm_max_ps(self.0, rhs.0)) }
+        unsafe { F32x4_41(_mm_max_ps(self.0, rhs.0)) }
     }
 
     #[inline(always)]
     fn min(self, rhs: Self) -> Self {
-        unsafe { F32x4(_mm_min_ps(self.0, rhs.0)) }
+        unsafe { F32x4_41(_mm_min_ps(self.0, rhs.0)) }
+    }
+}
+
+impl InternalSimdBaseIo for F32x4_41 {
+    #[inline(always)]
+    unsafe fn zeroes() -> Self {
+        F32x4_41(_mm_setzero_ps())
+    }
+
+    #[inline(always)]
+    unsafe fn set1(x: Self::Scalar) -> Self {
+        F32x4_41(_mm_set1_ps(x))
     }
 
     #[inline(always)]
     unsafe fn load_from_array(array: Self::ArrayRepresentation) -> Self {
-        F32x4(_mm_loadu_ps(array.as_ptr()))
+        F32x4_41(_mm_loadu_ps(array.as_ptr()))
     }
 
     #[inline(always)]
     unsafe fn load_from_ptr_unaligned(ptr: *const Self::Scalar) -> Self {
-        F32x4(_mm_loadu_ps(ptr as *const f32))
+        F32x4_41(_mm_loadu_ps(ptr as *const f32))
     }
 
     #[inline(always)]
@@ -773,7 +734,7 @@ impl SimdBase for F32x4 {
 
     #[inline(always)]
     unsafe fn load_from_ptr_aligned(ptr: *const Self::Scalar) -> Self {
-        F32x4(_mm_load_ps(ptr as *const f32))
+        F32x4_41(_mm_load_ps(ptr as *const f32))
     }
 
     #[inline(always)]
@@ -793,49 +754,33 @@ impl SimdBase for F32x4 {
 
     #[inline(always)]
     unsafe fn from_underlying_value(value: Self::UnderlyingType) -> Self {
-        F32x4(value)
+        F32x4_41(value)
     }
 }
 
-impl SimdFloat for F32x4 {
+impl SimdFloat for F32x4_41 {
     #[inline(always)]
     fn div(self, rhs: Self) -> Self {
-        unsafe { F32x4(_mm_div_ps(self.0, rhs.0)) }
+        unsafe { F32x4_41(_mm_div_ps(self.0, rhs.0)) }
     }
 
     #[inline(always)]
     fn ceil(self) -> Self {
-        unsafe {
-            Self::load_from_array([
-                self[0].m_ceil(),
-                self[1].m_ceil(),
-                self[2].m_ceil(),
-                self[3].m_ceil(),
-            ])
-        }
+        unsafe { F32x4_41(_mm_ceil_ps(self.0)) }
     }
 
     #[inline(always)]
     fn floor(self) -> Self {
-        unsafe {
-            Self::load_from_array([
-                self[0].m_floor(),
-                self[1].m_floor(),
-                self[2].m_floor(),
-                self[3].m_floor(),
-            ])
-        }
+        unsafe { F32x4_41(_mm_floor_ps(self.0)) }
     }
 
     #[inline(always)]
     fn round(self) -> Self {
         unsafe {
-            let sign_mask = F32x4::set1(-0.0);
-            let magic = I32x4::set1(0x4B000000).bitcast_f32();
-            let sign = self & sign_mask;
-            let signed_magic = magic | sign;
-            let b = self + signed_magic;
-            b - signed_magic
+            F32x4_41(_mm_round_ps(
+                self.0,
+                _MM_FROUND_TO_NEAREST_INT | _MM_FROUND_NO_EXC,
+            ))
         }
     }
 
@@ -877,97 +822,79 @@ impl SimdFloat for F32x4 {
     #[inline(always)]
     fn horizontal_add(self) -> Self::Scalar {
         unsafe {
-            let t1 = _mm_movehl_ps(self.0, self.0);
-            let t2 = _mm_add_ps(self.0, t1);
-            let t3 = _mm_shuffle_ps(t2, t2, 1);
-            _mm_cvtss_f32(t2) + _mm_cvtss_f32(t3)
+            let mut tmp = _mm_hadd_ps(self.0, self.0);
+            tmp = _mm_hadd_ps(tmp, tmp);
+            _mm_cvtss_f32(tmp)
         }
     }
 
     #[inline(always)]
     fn sqrt(self) -> Self {
-        unsafe { F32x4(_mm_sqrt_ps(self.0)) }
+        unsafe { F32x4_41(_mm_sqrt_ps(self.0)) }
     }
 
     #[inline(always)]
     fn rsqrt(self) -> Self {
-        unsafe { F32x4(_mm_rsqrt_ps(self.0)) }
+        unsafe { F32x4_41(_mm_rsqrt_ps(self.0)) }
     }
 }
 
-impl SimdFloat32 for F32x4 {
-    type SimdI32 = I32x4;
+impl SimdFloat32 for F32x4_41 {
+    type SimdI32 = I32x4_41;
 
     #[inline(always)]
     fn bitcast_i32(self) -> Self::SimdI32 {
-        unsafe { I32x4(_mm_castps_si128(self.0)) }
+        unsafe { I32x4_41(_mm_castps_si128(self.0)) }
     }
 
     #[inline(always)]
     fn cast_i32(self) -> Self::SimdI32 {
-        unsafe { I32x4(_mm_cvtps_epi32(self.0)) }
+        unsafe { I32x4_41(_mm_cvtps_epi32(self.0)) }
     }
 
     #[inline(always)]
     fn fast_inverse(self) -> Self {
-        unsafe { F32x4(_mm_rcp_ps(self.0)) }
+        unsafe { F32x4_41(_mm_rcp_ps(self.0)) }
     }
 }
 
-#[derive(Copy, Clone)]
-pub struct F64x2(__m128d);
-impl_simd_base_overloads!(F64x2);
-impl_simd_float_overloads!(F64x2);
+define_simd_type!(f64, 2, __m128d, _41);
+impl_simd_float_overloads!(F64x2_41);
 
-impl SimdBase for F64x2 {
-    const WIDTH: usize = 2;
-    type Scalar = f64;
-    type ArrayRepresentation = [f64; 2];
-    type UnderlyingType = __m128d;
-
-    #[inline(always)]
-    unsafe fn zeroes() -> Self {
-        F64x2(_mm_setzero_pd())
-    }
-
-    #[inline(always)]
-    unsafe fn set1(x: Self::Scalar) -> Self {
-        F64x2(_mm_set1_pd(x))
-    }
-
+impl SimdBaseOps for F64x2_41 {
     #[inline(always)]
     fn add(self, rhs: Self) -> Self {
-        unsafe { F64x2(_mm_add_pd(self.0, rhs.0)) }
+        unsafe { F64x2_41(_mm_add_pd(self.0, rhs.0)) }
     }
 
     #[inline(always)]
     fn sub(self, rhs: Self) -> Self {
-        unsafe { F64x2(_mm_sub_pd(self.0, rhs.0)) }
+        unsafe { F64x2_41(_mm_sub_pd(self.0, rhs.0)) }
     }
 
     #[inline(always)]
     fn mul(self, rhs: Self) -> Self {
-        unsafe { F64x2(_mm_mul_pd(self.0, rhs.0)) }
+        unsafe { F64x2_41(_mm_mul_pd(self.0, rhs.0)) }
     }
 
     #[inline(always)]
     fn bit_and(self, rhs: Self) -> Self {
-        unsafe { F64x2(_mm_and_pd(self.0, rhs.0)) }
+        unsafe { F64x2_41(_mm_and_pd(self.0, rhs.0)) }
     }
 
     #[inline(always)]
     fn bit_or(self, rhs: Self) -> Self {
-        unsafe { F64x2(_mm_or_pd(self.0, rhs.0)) }
+        unsafe { F64x2_41(_mm_or_pd(self.0, rhs.0)) }
     }
 
     #[inline(always)]
     fn bit_xor(self, rhs: Self) -> Self {
-        unsafe { F64x2(_mm_xor_pd(self.0, rhs.0)) }
+        unsafe { F64x2_41(_mm_xor_pd(self.0, rhs.0)) }
     }
 
     #[inline(always)]
     fn bit_not(self) -> Self {
-        unsafe { I64x2::set1(-1).bitcast_f64().bit_xor(self) }
+        unsafe { I64x2_41::set1(-1).bitcast_f64().bit_xor(self) }
     }
 
     #[inline(always)]
@@ -977,64 +904,74 @@ impl SimdBase for F64x2 {
 
     #[inline(always)]
     fn and_not(self, rhs: Self) -> Self {
-        unsafe { F64x2(_mm_andnot_pd(self.0, rhs.0)) }
+        unsafe { F64x2_41(_mm_andnot_pd(self.0, rhs.0)) }
     }
 
     #[inline(always)]
     fn blendv(self, a: Self, b: Self) -> Self {
-        self.bitcast_i64()
-            .blendv(a.bitcast_i64(), b.bitcast_i64())
-            .bitcast_f64()
+        unsafe { F64x2_41(_mm_blendv_pd(a.0, b.0, self.0)) }
     }
 
     #[inline(always)]
     fn cmp_eq(self, rhs: Self) -> Self {
-        unsafe { F64x2(_mm_cmpeq_pd(self.0, rhs.0)) }
+        unsafe { F64x2_41(_mm_cmpeq_pd(self.0, rhs.0)) }
     }
 
     #[inline(always)]
     fn cmp_neq(self, rhs: Self) -> Self {
-        unsafe { F64x2(_mm_cmpneq_pd(self.0, rhs.0)) }
+        unsafe { F64x2_41(_mm_cmpneq_pd(self.0, rhs.0)) }
     }
 
     #[inline(always)]
     fn cmp_lt(self, rhs: Self) -> Self {
-        unsafe { F64x2(_mm_cmplt_pd(self.0, rhs.0)) }
+        unsafe { F64x2_41(_mm_cmplt_pd(self.0, rhs.0)) }
     }
 
     #[inline(always)]
     fn cmp_lte(self, rhs: Self) -> Self {
-        unsafe { F64x2(_mm_cmple_pd(self.0, rhs.0)) }
+        unsafe { F64x2_41(_mm_cmple_pd(self.0, rhs.0)) }
     }
 
     #[inline(always)]
     fn cmp_gt(self, rhs: Self) -> Self {
-        unsafe { F64x2(_mm_cmpgt_pd(self.0, rhs.0)) }
+        unsafe { F64x2_41(_mm_cmpgt_pd(self.0, rhs.0)) }
     }
 
     #[inline(always)]
     fn cmp_gte(self, rhs: Self) -> Self {
-        unsafe { F64x2(_mm_cmpge_pd(self.0, rhs.0)) }
+        unsafe { F64x2_41(_mm_cmpge_pd(self.0, rhs.0)) }
     }
 
     #[inline(always)]
     fn max(self, rhs: Self) -> Self {
-        unsafe { F64x2(_mm_max_pd(self.0, rhs.0)) }
+        unsafe { F64x2_41(_mm_max_pd(self.0, rhs.0)) }
     }
 
     #[inline(always)]
     fn min(self, rhs: Self) -> Self {
-        unsafe { F64x2(_mm_min_pd(self.0, rhs.0)) }
+        unsafe { F64x2_41(_mm_min_pd(self.0, rhs.0)) }
+    }
+}
+
+impl InternalSimdBaseIo for F64x2_41 {
+    #[inline(always)]
+    unsafe fn zeroes() -> Self {
+        F64x2_41(_mm_setzero_pd())
+    }
+
+    #[inline(always)]
+    unsafe fn set1(x: Self::Scalar) -> Self {
+        F64x2_41(_mm_set1_pd(x))
     }
 
     #[inline(always)]
     unsafe fn load_from_array(array: Self::ArrayRepresentation) -> Self {
-        F64x2(_mm_loadu_pd(array.as_ptr()))
+        F64x2_41(_mm_loadu_pd(array.as_ptr()))
     }
 
     #[inline(always)]
     unsafe fn load_from_ptr_unaligned(ptr: *const Self::Scalar) -> Self {
-        F64x2(_mm_loadu_pd(ptr))
+        F64x2_41(_mm_loadu_pd(ptr))
     }
 
     #[inline(always)]
@@ -1044,7 +981,7 @@ impl SimdBase for F64x2 {
 
     #[inline(always)]
     unsafe fn load_from_ptr_aligned(ptr: *const Self::Scalar) -> Self {
-        F64x2(_mm_load_pd(ptr))
+        F64x2_41(_mm_load_pd(ptr))
     }
 
     #[inline(always)]
@@ -1064,37 +1001,33 @@ impl SimdBase for F64x2 {
 
     #[inline(always)]
     unsafe fn from_underlying_value(value: Self::UnderlyingType) -> Self {
-        F64x2(value)
+        F64x2_41(value)
     }
 }
 
-impl SimdFloat for F64x2 {
+impl SimdFloat for F64x2_41 {
     #[inline(always)]
     fn div(self, rhs: Self) -> Self {
-        unsafe { F64x2(_mm_div_pd(self.0, rhs.0)) }
+        unsafe { F64x2_41(_mm_div_pd(self.0, rhs.0)) }
     }
 
     #[inline(always)]
     fn ceil(self) -> Self {
-        unsafe { Self::load_from_array([self[0].m_ceil(), self[1].m_ceil()]) }
+        unsafe { F64x2_41(_mm_ceil_pd(self.0)) }
     }
 
     #[inline(always)]
     fn floor(self) -> Self {
-        unsafe { Self::load_from_array([self[0].m_floor(), self[1].m_floor()]) }
+        unsafe { F64x2_41(_mm_floor_pd(self.0)) }
     }
 
     #[inline(always)]
     fn round(self) -> Self {
         unsafe {
-            let sign_mask = Self::set1(-0.0);
-            let magic = Self(_mm_castsi128_pd(_mm_set_epi32(
-                0x43300000, 0, 0x43300000, 0,
-            )));
-            let sign = self & sign_mask;
-            let signedmagic = magic | sign;
-            let b = self + signedmagic;
-            b - signedmagic
+            F64x2_41(_mm_round_pd(
+                self.0,
+                _MM_FROUND_TO_NEAREST_INT | _MM_FROUND_NO_EXC,
+            ))
         }
     }
 
@@ -1143,7 +1076,7 @@ impl SimdFloat for F64x2 {
 
     #[inline(always)]
     fn sqrt(self) -> Self {
-        unsafe { F64x2(_mm_sqrt_pd(self.0)) }
+        unsafe { F64x2_41(_mm_sqrt_pd(self.0)) }
     }
 
     #[inline(always)]
@@ -1152,12 +1085,12 @@ impl SimdFloat for F64x2 {
     }
 }
 
-impl SimdFloat64 for F64x2 {
-    type SimdI64 = I64x2;
+impl SimdFloat64 for F64x2_41 {
+    type SimdI64 = I64x2_41;
 
     #[inline(always)]
     fn bitcast_i64(self) -> Self::SimdI64 {
-        unsafe { I64x2(_mm_castpd_si128(self.0)) }
+        unsafe { I64x2_41(_mm_castpd_si128(self.0)) }
     }
 
     #[inline(always)]
