+++ src/tests/run.rs
@@ -2,41 +2,42 @@ use crate::{elementwise_eq_tester, test_constify_imm8};
 
 use super::*;
 
-use crate::{avx2::*, scalar::*, sse2::*, sse41::*, *};
+use crate::engines::{avx2::*, scalar::*, sse2::*, sse41::*};
+use crate::*;
 
-elementwise_eq_tester_impl!(SimdBase, add, two_arg, EqPrecision::exact());
-elementwise_eq_tester_impl!(SimdBase, sub, two_arg, EqPrecision::exact());
-elementwise_eq_tester_impl!(SimdBase, mul, two_arg, EqPrecision::exact());
+elementwise_eq_tester_impl!(SimdBaseOps, add, two_arg, EqPrecision::exact());
+elementwise_eq_tester_impl!(SimdBaseOps, sub, two_arg, EqPrecision::exact());
+elementwise_eq_tester_impl!(SimdBaseOps, mul, two_arg, EqPrecision::exact());
 
-elementwise_eq_tester_impl!(SimdBase, bit_and, two_arg, EqPrecision::exact());
-elementwise_eq_tester_impl!(SimdBase, bit_or, two_arg, EqPrecision::exact());
-elementwise_eq_tester_impl!(SimdBase, bit_xor, two_arg, EqPrecision::exact());
-elementwise_eq_tester_impl!(SimdBase, bit_not, one_arg, EqPrecision::exact());
-elementwise_eq_tester_impl!(SimdBase, and_not, two_arg, EqPrecision::exact());
+elementwise_eq_tester_impl!(SimdBaseOps, bit_and, two_arg, EqPrecision::exact());
+elementwise_eq_tester_impl!(SimdBaseOps, bit_or, two_arg, EqPrecision::exact());
+elementwise_eq_tester_impl!(SimdBaseOps, bit_xor, two_arg, EqPrecision::exact());
+elementwise_eq_tester_impl!(SimdBaseOps, bit_not, one_arg, EqPrecision::exact());
+elementwise_eq_tester_impl!(SimdBaseOps, and_not, two_arg, EqPrecision::exact());
 
 // Abs breaks on minimum integer values as they dont have a corrisponding maximum, causing undefined behavior.
-elementwise_eq_tester_impl!(SimdBase, abs, one_arg_abs_filtered, EqPrecision::exact());
+elementwise_eq_tester_impl!(SimdBaseOps, abs, one_arg_abs_filtered, EqPrecision::exact());
 
-elementwise_eq_tester_impl!(SimdBase, cmp_eq, two_arg, EqPrecision::exact());
-elementwise_eq_tester_impl!(SimdBase, cmp_lt, two_arg, EqPrecision::exact());
-elementwise_eq_tester_impl!(SimdBase, cmp_lte, two_arg, EqPrecision::exact());
-elementwise_eq_tester_impl!(SimdBase, cmp_gt, two_arg, EqPrecision::exact());
-elementwise_eq_tester_impl!(SimdBase, cmp_gte, two_arg, EqPrecision::exact());
+elementwise_eq_tester_impl!(SimdBaseOps, cmp_eq, two_arg, EqPrecision::exact());
+elementwise_eq_tester_impl!(SimdBaseOps, cmp_lt, two_arg, EqPrecision::exact());
+elementwise_eq_tester_impl!(SimdBaseOps, cmp_lte, two_arg, EqPrecision::exact());
+elementwise_eq_tester_impl!(SimdBaseOps, cmp_gt, two_arg, EqPrecision::exact());
+elementwise_eq_tester_impl!(SimdBaseOps, cmp_gte, two_arg, EqPrecision::exact());
 
-elementwise_eq_tester_impl!(SimdBase, blendv, iter_blendv_ags, EqPrecision::exact());
+elementwise_eq_tester_impl!(SimdBaseOps, blendv, iter_blendv_ags, EqPrecision::exact());
 
 // We filter out NaN numbers for neq because under some hardware implementations (including Avx2)
 // it appears that `NaN != [number]` is false. Technically that's invalid according to the floatin point
 // spec, but it seems like a hardware thing that we cant avoid.
 elementwise_eq_tester_impl!(
-    SimdBase,
+    SimdBaseOps,
     cmp_neq,
     two_arg_nan_filtered,
     EqPrecision::exact()
 );
 
-elementwise_eq_tester_impl!(SimdBase, max, two_arg_nan_filtered, EqPrecision::exact());
-elementwise_eq_tester_impl!(SimdBase, min, two_arg_nan_filtered, EqPrecision::exact());
+elementwise_eq_tester_impl!(SimdBaseOps, max, two_arg_nan_filtered, EqPrecision::exact());
+elementwise_eq_tester_impl!(SimdBaseOps, min, two_arg_nan_filtered, EqPrecision::exact());
 
 // We use "almost" precision for these functions because at higher numbers there's always small errors.
 // However if there's an error in smaller numbers, it would be caught pretty easily by this precision.
